+++ main.rkt
@@ -2,7 +2,12 @@
 (require "color.rkt" "private/core.rkt"
          (only-in "private/native.rkt"
                   skia-check! skia-available? skia-native-version
-                  skia-native-library-path native-package-version))
+                  skia-native-library-path native-package-version)
+         (only-in "private/harfbuzz-native.rkt"
+                  harfbuzz-check! harfbuzz-available? harfbuzz-native-version
+                  harfbuzz-native-library-path harfbuzz-package-version))
 (provide (all-from-out "color.rkt" "private/core.rkt")
          skia-check! skia-available? skia-native-version
-         skia-native-library-path native-package-version)
+         skia-native-library-path native-package-version
+         harfbuzz-check! harfbuzz-available? harfbuzz-native-version
+         harfbuzz-native-library-path harfbuzz-package-version)